Better validation and testing of key paths
Previously, we allowed key paths to end in backslashes and we allowed key
names to be more than 255 characters, which are both wrong.

This commit uses a regex to perform better validation. It also adds
better tests, and to ensure that the capture of the repeating capture
group works (for more than a single level).