Encode default-ness in the registry path
Previously, there was a separate parameter for specifying whether a
registry value is the unnamed (aka default). However, the following
example resulted in non unique paths:

  registry_value { 'hklm\foo':
    isdefault => true
  }
  registry_value { 'hklm\foo':
    isdefault => false
  }

The first value is managing the default value for the `hklm\foo` key. The
second value is managing the value named `foo` for the `hklm` key.

This commit eliminates this issue and eliminates the need for an isdefault
parameter. The above examples become:

  registry_value { 'hklm\foo\\':
  }
  registry_value { 'hklm\foo':
  }

Note, the trailing backslash must be escaped, since it is immediately
followed by a single quote.