Fix autorequiring when using different root key forms
Previously, autorequiring would not add relationships when resources mix
short and long forms of the root key, e.g. hklm and hkey_local_machine. It
also did not handle different cases, hklm vs HKLM.

This commit ensures that we always create a canonical path for the
`KeyPath` and `ValuePath`, which always use the downcased root key name.
In the process, the ascend logic was greatly simplified.

This commit does not fix general case-insensitivity in key and value
names, e.g. so hklm\Software and hklm\software are still seen as different
resources. This will be addressed in a separate commit.