Handle binary registry values
This commit adds support for binary registry values. They must be hex
encoded strings with the high nibble first. The strings can be space
delimited, as is done in Microsoft's reg format, or omitted. So:

   deadbeef
   DE AD BE EF

Are both valid and equivalent. Empty string is also permitted.

This commit also adds better validation for dword and qword. The former is an
unsigned 32-bit integer, the latter unsigned 64-bit integer.