Add examples of current registry key and value types
Without this patch the examples of the registry key and value types and
providers are not up to date.  This is a problem because it can be hard
to quickly understand what the end user perspective and experience is
meant to be.

This patch fixes the problem by providing an example manifest with up to
date examples written with the perspective of the end user in mind.