Add REG_MULTI_SZ (type => array) implementation
Without this patch applied the registry type and provider do not have an
implementation for the following resource declaration.

    registry_value { 'HKLM\Software\Vendor\Bar\value_array':
      ensure => present,
      type   => array,
      data   => [ 'one', 'two', 'three' ],
    }

The `type => array` in the Puppet DSL maps to the REG_MULTI_SZ [1] data type.
This is basically an ordered list of null terminated strings with two null
bytes to denote the end of the list.

In Puppet we've implemented this as simply accepting an array for the data
property.  This implementation also works well with `puppet resource`:

    PS C:\> puppet resource registry_value HKLM\Software\Vendor\Bar\somearray
    registry_value { 'HKLM\Software\Vendor\Bar\somearray':
      ensure => 'present',
      data   => ['one', 'two', 'three'],
      type   => 'array',
    }

The change_to_s method on the data property has also been overriden to
display messages in the form of:

    data changed 'one,three,two,four' to 'one,two,three'

[1] http://msdn.microsoft.com/en-us/library/windows/desktop/ms724884.aspx