Allow values associated with a registry key to be purged
Previously, puppet could only manage registry values as a 'minimal' set,
but not as an 'inclusive' set.

This commit adds a `purge` parameter to the registry_key type. If set to
true, puppet will purge all registry values associated with the key that
are not being managed by puppet. For example, if a key `foo` contains
values `bar`, `baz`, and puppet is only managing `baz` and `qux`, then
`bar` would be purged (deleted).

  registry_key { 'hklm\software\foo':
    ensure => present,
    purge  => true
  }
  registry_value { 'hklm\software\baz':
    ensure => present,
    type   => string,
    data   => 'val1'
  }
  registry_value { 'hklm\software\qux':
    ensure => present,
    type   => string,
    data   => 'val2'
  }

This commit satisfies the use-case where the user wants to specify the
exact set of values for a key and ensure that no other values exist
(without having to know what their names might be).