Work around #3947, #4248, #14073; load our utility code
Without this patch the registry module blows up when installed onto a
Linux Puppet Master.  The utility code and the structure of the types
and providers end up requiring win32/registry which will never be
present on non-windows systems.  In addition, the types and providers
mix in module level support code, which does not work because of the
following Puppet bugs, all related to our inability to load utility code
from the $LOAD_PATH.

 * https://projects.puppetlabs.com/issues/3947
 * https://projects.puppetlabs.com/issues/4248
 * https://projects.puppetlabs.com/issues/14073

This has annoyed me enough that I'm going revisit these bugs and the
interplay of the modulepath, environments, and $LOAD_PATH.