Add registry::service defined resource example
Without this patch we don't have any easy to use examples of how to use
the registry types and providers.

This patch fixes the problem by including a defined resource type named
registry::service and an example class named registry::service_example.
After installing the registry module, the end user need only include
this class and the appropriate registry keys and values will be managed.

The current implementation requires a reboot for the services created by
these registry entries to show up in the Service Control Manager
(`services.msc`).

An example of how to use this new defined resource is:

      registry::service { 'PuppetExample1':
        display_name => "Puppet Example 1",
        description  => "Puppet Example Service 1",
        command      => 'C:\PuppetExample1.bat',
        start        => 'disabled',
      }