Allow values associated with a registry key to be purged
Previously, puppet could only manage registry values as a 'minimal' set,
but not as an 'inclusive' set.

This commit adds a `purge_values` parameter to the registry_key type. If
set to true, puppet will purge all registry values associated with the
key that are not being managed by puppet. For example, if a key `foo`
contains values `bar`, `baz`, and puppet is only managing `baz` and
`qux`, then `bar` would be purged (deleted).

This patch does not purge registry keys that are children of a registry
key resource being managed.  It only purges values.

  registry_key { 'hklm\software\foo':
    ensure       => present,
    purge_values => true
  }
  registry_value { 'hklm\software\baz':
    ensure => present,
    type   => string,
    data   => 'val1'
  }
  registry_value { 'hklm\software\qux':
    ensure => present,
    type   => string,
    data   => 'val2'
  }

This commit satisfies the use-case where the user wants to specify the
exact set of values for a key and ensure that no other values exist
(without having to know what their names might be).

Add registry::purge_example class

This patch adds a class to the registry module to make it easier for the
user to see the behavior of the purge_values registry_key property.

The user needs to include the registry::purge_example class in the node
catalog, then set the following environment variables:

    PS C:\> $env:FACTER_PURGE_EXAMPLE_MODE = 'setup'
    PS C:\> puppet agent --test

This command will configure a registry key with a sub key and values
associated with the key.

    PS C:\> $env:FACTER_PURGE_EXAMPLE_MODE = 'purge'
    PS C:\> puppet agent --test

This command will configure a registry key with purge_values => true and
manage only half of the registry_value resources managed in the setup
phase.