Separate the implementation of the type and provider
Without this patch, the Win32 library and API are tangled up in both the
type and provider.  The type should not have any Win32 specific logic,
as this code is meant to be run on the Puppet Master and compiler nodes.

This patch fixes the problem by separating out helper methods into two
separate modules.

This patch introduces two new classes, RegistryValuePath and
RegistryKeyPath.  These two classes have slightly different semantics
for what a subkey is.  These classes are strictly meant to perform
manipulation of a registry path string.

This patch separates out the Win32 API into
Puppet::Modules::Registry::ProviderBase which implements instance
methods meant to be mixed into the provider classes.

The patch removes all of the Win32 API logic hanging off of the path
parameter instance itself.  All of this logic is moved into instance
methods of the provider itself instead of the parameter.  This is
necessary to separate out the behavior from the type instances itself,
which share the parameter instances.

Without this patch the spec tests do not pass on a Unix system.  Since
all of the Win32 specific logic is removed from the type by this patch
the spec tests now pass on Unix systems.