Add acceptance tests for the registry type
Without this patch there are no acceptance tests for the Puppet Labs
registry module.  This is a problem because we need to ensure the
registry_key and registry_value providers actually manage the registry
as expected.

This patch partially fixes the problem by adding a single acceptance
test.  This is a partial fix because I want to lay all of the groundwork
for additional acceptance tests first, then add additional tests in
subsequent patches.

This patch introduces a Ruby module named Systest::Util::Registry.  This
module is meant to be mixed into the eigenclass of a specific test case.
All of the helper methods will then be available to the test case
instance.

A single test case example is provided to help illustrate how to add
additional acceptance tests.  Please copy registry/should_pluginsync.rb
to additional files to create additional tests.