Fix cannot alias error when managing 32 and 64 bit versions of a key
Without this patch applied the Registry module is not able to manage
both the 32 and 64 bit copies of the same-named registry subkey.

The acceptance test accompanied by this patch illustrates this failure.
Without the change to the way keys are canonically named with the path
namevar the following error message is shown:

    err: Failed to apply catalog: Cannot alias \
      Registry_key[32:HKLM\Software\Vendor\PuppetLabsTest_tYEtHVZl\SubKey1]
      to
      ["hklm\\Software\\Vendor\\PuppetLabsTest_tYEtHVZl\\SubKey1"]

This patch fixes the error by playing the 32: substring prefix in the
canonical string path returned by the path parameter munging block.

TODO: Add a spec test.