Add Registry_key ensure => absent and purge_values coverage
This patch adds test coverage for the following two behaviors of the
Registry_key type:

 * The purge_values => true parameter.
 * The ensure => absent parameter.

The patch implements this by running a series of Puppet agent runs with
the same Puppet Master running.  The behavior is dependent on the
FACT_PHASE facter fact set by the harness using the environment option.