Add acceptance tests for registry_value provider
Without this patch there is extremely limited acceptance test coverage
for the `registry_value` provider.

This patch adds a new test case which will be a comprehensive place to
test the behavior of the registry_value provider.

The test coverage provided by this patch exercises the following things:

Create registry_value resources with the following data types:

 * string (REG_SZ)
 * expand (REG_EXPAND_SZ)
 * array  (REG_MULTI_SZ)
 * dword  (REG_DWORD)
 * qword  (REG_QWORD)
 * binary (REG_BINARY)

This list is comprehensive against what the registry_value type supports
at the time of this patch.

The tests will then re-declare all of these resources used to create
these values, only using slightly different data.  This exercises the
transaction aspect of the RAL, ensuring we properly display change
methods.

At each step of create, change, we run puppet a second time and ensure
the expected create or change methods are not displayed.  This is our
assurance Puppet is idempotent and the resources were actually properly
managed to begin with.