Eliminate RegistryPathBas#(default?,valuename) from base class
Josh pointed out that the current implementation doesn't actually use
the RegistryPathBase#default? and RegistryPathBase#valuename instance
methods.  These methods are specific to the implementation of the
RegistryKeyPath or RegistryValuePath instances.

This patch simply removes the methods from the base class and adds them
explicitly to the ancestors to make it clear where the method being
called is defined any why it's there.