(maint) Remove RegistryKeyPath#{valuename,default?} methods
The system should never need to call the valuename or default? methods
on a RegistryKeyPath instance.  Instead, only RegistryValuePath
instances should respond to the valuename and default? methods.

This patch simply removes the methods from the RegistryKeyPath class to
make it clear these methods are not available or meaningful when dealing
with key paths.