(#14501) Fix autorequire case sensitivity
Without this patch the behavior of a registry_value auto requiring
parent registry_key resources does not work when the parent resources do
not share the same letter case as the value.

For example, the value fails to require the key in this scenario:

    registry_key { 'HKLM\Software\Vendor\PUPPETLABS':
      ensure => present,
    }
    registry_value { 'HKLM\Software\Vendor\puppetlabs\ValueString':
      data => "Hello World!",
    }

The root cause of this problem is that Puppet treats all resource
identifier keys as case sensitive and the registry types ascend their
`path` attribute looking up resources using the string provided by the
user in the path parameter.

This patch fixes the problem by aliasing every registry_key and
registry_value resource using the downcase version of the path
attribute.  Other resources are expected to look up related resources in
the catalog using the downcase version of the path parameter.

This solves the problem because we now have an internally consistent,
case insensitive identifier.

We preserve case for the user by not munging the case of the path
parameter itself.  Instead we simply alias the downcase version in the
catalog.