(Maint) Add type documentation
Previously, running `puppet resource registry_key` or `puppet resource
registry_value` only contained documentation for the parmeters, but not
the types themselves.

Also, some of the parameter documentation lines were too long and
wrapping, but then breaking after the first word on the next line.

This commit adds documentation for the types and updates formatting
for some of the parameters.