(#14529) Add registry::value defined type
Without this patch we don't have an easy to use way to abstract away the
management of a value and the key it's in.

This patch fixes the problem by creating a new defined resource type in
Puppet that builds on top of the native `registry_key` and
`registry_value` resource types.

This registry::value type allows the user to manage values without
worrying about also managing the key the value exists in.  For example:

    registry::value { 'MyAPP - Owner Email':
      key   => 'HKLM\Software\Puppet Labs',
      value => 'Owner Email',
      data  => 'jeff@puppetlabs.com',
    }

This resource will automatically manage the parent key if it does not
exist and manage the value.  The key and value will be created in the
native architecture of the hive.  That is to say, the 64 bit hive on a
64 bit system and the 32 bit hive on a 32 bit system.