(#14572) Fix management of the default value
Without this patch applied the default value of a registry key is not
properly managed by Puppet.

The root cause is that the munge operation of the registry_value type
causes the trailing slash to be stripped off when stored in the
parameter value of the resource.

This, in turn, caused the provider to behave as if it was managing a
normal value rather than a default value.

This patch fixes the problem by changing the munge method.  Munging the
path of a default value preserves the trailing slash with this patch
applied.  This causes the provider to treat the value as a default value
rather than a normal value.

A new method, RegistryPath#aliases returns an array of alases a given
Key or Value resource should create in the catalog given a path.  This
makes it easier to match up aliases created in the munge block with
aliases referenced in the autorequire block.

Finally, the type of path used in the autorequire block is changed from
RegistryKeyPath to RegistryValuePath.  Without this change a default
value would skip over the parent key since the instance parsed it as a
path and not a key.

The spec tests are updated by this patch to pass a catalog instance to
the initializer whenever creating an instance of
Puppet::Type.type(:registry_{value,key})  This is required when
declaring aliases from the munge method.