(#14517) Improve error handling when writing values
Without this patch applied if Puppet tries to write a registry value
into a parent key that does not exist the following error message is
returned to the user:

    Could not evaluate: The system cannot find the file specified.

This is a poor error message because we're not talking about files but
instead registry keys.

This patch fixes the problem by catching the exception, testing the
message against "cannot find" and augmenting the error message to be
more helpful before re-raising it.

In the event we don't know why the exception was raised, we augment the
message with "Unexpected exception from Win32 API..." and include the
detail.