(#14149) Refactor utility code into PuppetX namespace
Without this patch the registry module is a bad example for contributors
and module authors because it does not follow the PuppetX convention
described at Puppet Design Guidelines [1]

This patch fixes the problem by moving all of the utility code and
mix-in modules from Puppet::Modules::Registry to
PuppetX::Puppetlabs::Registry namespace.

In addition, the loading code (require statements) have been modified to
assume #7788 (rubygems support) is implemented.  In the scenario where
7788 is not yet implemented we fall back to the work around of requiring
fully qualified paths to operate with pluginsync as described in #14073.

[1] http://projects.puppetlabs.com/projects/puppet/wiki/Puppet_Design_Guidelines