(Maint) Fix unhandled exception working around #4248
Without this patch applied the Puppet master will error out because of
an uncaught exception trying to load utility code.  This is a problem
because the registry types and providers cause catalog compilation to
fail.

The root cause of this problem is that require without an exception
class will default to a behavior that does not catch LoadError
exceptions.

This patch fixes the problem by explicitly catching LoadError and
re-raising if the work-around fails.