Updated content to conform to README best practices template
Before alterations, this content was the module author- determined
description of and instructions for use of the module.

As part of a joint Forge/Docs team effort to standardize formatting and
encourage quality module documentation, a best practices README
template was created via internal and external user testing. That
template was then applied to this module.

I pulled in content from the original README on GitHub and in-code
comments. Standard headings were added (Overview, Module Description,
Setup, Usage, etc.) to organize content, existent content was moved
under its appropriate heading and edited for tone/flow/clarity, and
basic formatting was done to adhere to template standards.
I chose to leave off a Reference section as all bits of the inner
workings of the module are called out and discussed under Usage.