Only try to manage redirected registry keys/values on x64 systems
Previously, the acceptance test was trying to manage redirected keys and
values on x86 and x64 systems. However, on x86 systems, both 32:HKLM and
HKLM refer to the same key. As a result, only one 'created' message was
output for each pair of key/values, causing the acceptance test to fail.

This commit ensures we only test the redirected keys/values, those
prefixed with '32:' when run on x64 systems.