Detect whether sysnative filesystem alias is available
Previously the test assumed that the sysnative filesystem alias was
available, so that we could run the 64-bit version of reg.exe. However,
this fails when running on x86 versions of Windows. It also fails when
running on x64 on 2003, as the sysnative alias is not available until
Vista and later.

This commit logs a warning if the sysnative alias is not available on
x64 systems. Otherwise, it uses the 'native' version of reg.exe to
check for the default registry value.