Refactor class variable usage in registry_value
Without this commit the provider for the registry_value type was using a
class instance variable. The purpose of this variable was to share a
single instance of a `Win32API` object between all provider instances.
The method of accessing this shared object from the provider instances
was to call the class variable with '@@' prepended. Starting with Ruby 1.9,
use of this method is considered harmful and now causes a warning to be rasied.

This commit adds an accessor class method to replace accessing the class
instance variable directly. The call that originally accessed the class
variable has been refactored to use the accessor method of the class.