maint: Add Gemfile
Without this someone may not have everything installed. With this they would
not be required to install puppet, facter, or hiera as long as they have the
proper locations for the source set. This is useful because having facter
installed as a gem when you are trying to work on the source can be difficult.