(FM-4917) Rake / Beaker / Rspec compatibility
Beaker needs a version of Rake less than 11 due to using
`last_comment' in rake_task.rb:62. Reduce the version of
Rake down to the same dependency that Beaker does as a
development dependency.

Rspec 3+ requires that should be explicitly set as an expect term.
be_true and be_false have also been changed so that they are not
whether something is truthy or falsey, but simply really true/false. To
get the older behavior, the test simply needs to check for be_truthy or
be_falsey. For explicit true/false checks, it is be true and be false.

Remove gem install bundler as part of modulesync updates.

Paired with Glenn Sarti <glenn.sarti at puppetlabs dot com>