(MODULES-3865) Manage keys in HKU (hkey_users)
Previously, the registry module only allowed managing keys and
values in HKLM and HKCR. This means the module would not be able to
manage keys in HKU\.DEFAULT, a well-known key store. Allow HKU keys
and values to also be managed with the module.

NOTE: This is not the same as managing a user hive and registry
keys. It is to purely manage the keys and values in HKU.