(MODULES-2957) Adding the ability to have backslashes in registry
Previously the registry_value resource could not update values with a backslash
in their name.  This was due to using the path as a way of expressing a unique
registry value, however values can only be uniquely named using a path and name.

This means the registry_value type needed a new parameter called `value_name`
which is a namevar.  Fortunately registry keys can not have a backslash in their
name so title_patterns could be used to split a resource title into it's path
and value name components.  This meant the type could maintain a level of
backwards compatibility.

This meant however that many of the assumptions about the :path parameter were
no longer true;
* It no longer contained the registry value_name (that is now in a separate
  param) so canonicalization is now the same as a registry_key
* Trailing backslashes needed to be deleted instead of kept
* Almost all of the custom code for the RegistryValuePath type was now redundant
* In most of the test cases, :path was used but this is no longer enough to
  identify a unique reg key value (It needs a Path and Value), however this
  could simply be changed to :title and then the title_pattern matching would
  extract the path and value_name.  This could also verify the backwards
  compatibiltiy behaviour of using a title

This commit also updates the unit tests for the new value_name parameter.

Unfortunately this is a breaking change even with the title_patterns.  Consider
the following:
  registry_value { 'atitle\value':
    path => 'HKLM\Software\foo',
  }
Previously this would manage the value called 'foo' in the key 'HKLM\Software'.
With this change it would manage the value called 'value' (extracted from the
title_pattern because value_name is not specified) in the key 'HKLM\Software\foo'

  registry_value { 'atitle':
    path => 'HKLM\Software\foo',
  }
Previously this would manage the value called 'foo' in the key 'HKLM\Software'.
With this change it would manage the default value (extracted from the
title_pattern because value_name is not specified) in the key 'HKLM\Software\foo'

This commit also reorganises the README slightly as previously the information
was out of order, and hard to decipher.