(MODULES-4446) Modify README and update the registry_value type
Previously the registry_value type stated that it created parent keys
if they did not exist, however this is not true.  This commit updates the text
to indicate that if the parent keys do not exist an error will raised, and a
recommendation to use the registry_key to ensure the parent keys exist first.

This commit also reformats the markdown so that there is a single space between
the header hashes and the text as per Puppet markdown guidelines.

[skip ci]