(MODULES-4976) Remove rspec configuration for win32_console
Previously the spec_helper would configure rspec to output all to STDOUT due to
issues with the win32_console gem.  However as that gem was removed in Puppet 4,
it is no longer required.  Also by redirecting to stdout, when using the junit
formatter, the output is sent to STDOUT instead of the specificed text file.
This commit removes the redundant rpsec configuration.