(MODULES-2957) Refactor provider base as module
Previously the ProviderBase was a mixin module that provided only a few common
methods.  However due to how puppet resource and puppet apply instantiates the
object, some methods could be missing.  This commit refactors the ProviderBase
from a mixin back to "vanilla" method calls which removes the instantiation
issue.  This also moves the FFI definition into registry_value as it is not
shared in registry_key.  While this does result in slightly more lines of code,
it is far more readable with much less indirection.