(maint) Explicitly require types in define tests
Previously the spec/define/value_spec.rb was failing when run by itself, however
when run as part of a group of tests it would pass.  This was due to the type
loading not running first, which then populated the needed Win32 namespaces on
non-Windows platforms.  Without this change Travis CI would intermittently fail
and spec tests could only be run together in order to pass.