(maint) Remove default for dword, qword data
This property previously set a default value of empty string, or ''.

An empty string is not a valid dword or qword. This commit modifies the
data property to not set a default value when the type is dword or
qword.

We probably shouldn't ever set a default value for data. For now though,
by limiting the change to dword and qword types, we should avoid making
any functional changes.